+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Enhancements
 
   * [Alixir.SMS.Env] support get env variable through process dictionary, support envs: `region_id`, `access_key_id` and `access_secret`
+  * [Alixir.SMS] support send_sms with options